+++ src/components/activity-page/managed-interactive/iframe-runtime.tsx
@@ -1,7 +1,8 @@
 // cf. https://github.com/concord-consortium/question-interactives/blob/master/src/scaffolded-question/components/iframe-runtime.tsx
-import React, { useRef } from "react";
-import { ILinkedInteractive } from "@concord-consortium/lara-interactive-api";
-import { useInteractivePhoneWrapper } from "./use-interactive-phone-wrapper";
+import React, { useEffect, useRef, useState } from "react";
+import { IframePhone } from "../../../types";
+import iframePhone from "iframe-phone";
+import { IGetFirebaseJwtRequest, ILinkedInteractive } from "@concord-consortium/lara-interactive-api";
 
 const kDefaultHeight = 300;
 
@@ -21,19 +22,66 @@ interface IProps {
   proposedHeight?: number;
   containerWidth?: number;
   setNewHint: (newHint: string) => void;
+  handleGetFirebaseJWT: (phone: IframePhone, request: IGetFirebaseJwtRequest) => void;
 }
 
 export const IframeRuntime: React.FC<IProps> =
   ({ url, authoredState, initialInteractiveState, setInteractiveState, linkedInteractives,
-      report, proposedHeight, containerWidth, setNewHint }) => {
-  const setInteractiveStateRef = useRef<(state: any) => void>(setInteractiveState);
+      report, proposedHeight, containerWidth, setNewHint, handleGetFirebaseJWT }) => {
+  const [ heightFromInteractive, setHeightFromInteractive ] = useState(0);
+  const [ ARFromSupportedFeatures, setARFromSupportedFeatures ] = useState(0);
+  const iframeRef = useRef<HTMLIFrameElement>(null);
+  const phoneRef = useRef<IframePhone>();
+  const setInteractiveStateRef = useRef<((state: any) => void)>(setInteractiveState);
   setInteractiveStateRef.current = setInteractiveState;
+  const linkedInteractivesRef = useRef(linkedInteractives?.length ? { linkedInteractives } : undefined);
 
-  const { iframeRef, height: heightFromInteractive, aspectRatio } = useInteractivePhoneWrapper({
-    url, report, authoredState, linkedInteractives, initialInteractiveState,
-    setInteractiveState, setHint: setNewHint });
+  useEffect(() => {
+    const initInteractive = () => {
+      const phone = phoneRef.current;
+      if (!phone) {
+        return;
+      }
+      phone.addListener("interactiveState", (newInteractiveState: any) => {
+        setInteractiveStateRef.current(newInteractiveState);
+      });
+      phone.addListener("height", (newHeight: number) => {
+        setHeightFromInteractive(newHeight);
+      });
+      phone.addListener("supportedFeatures", (info: any) => {
+        if (info.features.aspectRatio) {
+          setARFromSupportedFeatures(info.features.aspectRatio);
+        }
+      });
+      phone.addListener("getFirebaseJWT", (request: IGetFirebaseJwtRequest) => {
+        handleGetFirebaseJWT(phone, request);
+      });
+      phone.addListener("hint", (newHint: any) => {
+        setNewHint(newHint.text || "");
+      });
+      phone.post("initInteractive", {
+        mode: report ? "report" : "runtime",
+        authoredState,
+        interactiveState: initialInteractiveState,
+        ...linkedInteractivesRef.current
+      });
+    };
 
-  const heightFromSupportedFeatures = aspectRatio && containerWidth ? containerWidth / aspectRatio : 0;
+    if (iframeRef.current) {
+      // Reload the iframe.
+      iframeRef.current.src = url;
+      // Re-init interactive, this time using a new mode (report or runtime).
+      phoneRef.current = new iframePhone.ParentEndpoint(iframeRef.current, initInteractive);
+    }
+    // Cleanup.
+    return () => {
+      if (phoneRef.current) {
+        phoneRef.current.disconnect();
+      }
+    };
+  }, [url, authoredState, report, initialInteractiveState, setNewHint, handleGetFirebaseJWT]);
+
+  const heightFromSupportedFeatures = ARFromSupportedFeatures && containerWidth ? containerWidth / ARFromSupportedFeatures : 0;
   // There are several options for specifying the iframe height. Check if we have height specified by interactive (from IframePhone
   // "height" listener), height based on aspect ratio specified by interactive (from IframePhone "supportedFeatures" listener),
   // or height from container dimensions and embeddable specifications.
